pretty alerts working

+++ src/components/BookingCard.js
@@ -1,132 +1,3 @@
-// import React from "react";
-// import { makeStyles } from "@material-ui/core/styles";
-// import clsx from "clsx";
-// import Card from "@material-ui/core/Card";
-// import CardHeader from "@material-ui/core/CardHeader";
-// import CardMedia from "@material-ui/core/CardMedia";
-// import CardContent from "@material-ui/core/CardContent";
-// import CardActions from "@material-ui/core/CardActions";
-// import Collapse from "@material-ui/core/Collapse";
-// import Avatar from "@material-ui/core/Avatar";
-// import IconButton from "@material-ui/core/IconButton";
-// import Typography from "@material-ui/core/Typography";
-// import { red } from "@material-ui/core/colors";
-// import FavoriteIcon from "@material-ui/icons/Favorite";
-// import ShareIcon from "@material-ui/icons/Share";
-// import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-// import MoreVertIcon from "@material-ui/icons/MoreVert";
-
-// const useStyles = makeStyles(theme => ({
-//   card: {
-//     maxWidth: 345
-//   },
-//   media: {
-//     height: 0,
-//     paddingTop: "56.25%" // 16:9
-//   },
-//   expand: {
-//     transform: "rotate(0deg)",
-//     marginLeft: "auto",
-//     transition: theme.transitions.create("transform", {
-//       duration: theme.transitions.duration.shortest
-//     })
-//   },
-//   expandOpen: {
-//     transform: "rotate(180deg)"
-//   },
-//   avatar: {
-//     backgroundColor: red[500]
-//   }
-// }));
-
-// export default function BookingCard() {
-//   const classes = useStyles();
-//   const [expanded, setExpanded] = React.useState(false);
-
-//   function handleExpandClick() {
-//     setExpanded(!expanded);
-//   }
-
-//   return (
-//     <Card className={classes.card}>
-//       <CardHeader
-//         avatar={
-//           <Avatar aria-label="Recipe" className={classes.avatar}>
-//             Booking
-//           </Avatar>
-//         }
-//         action={
-//           <IconButton aria-label="Settings">
-//             <MoreVertIcon />
-//           </IconButton>
-//         }
-//         title="Booking1"
-//         subheader="September 14, 2016"
-//       />
-//       <CardMedia
-//         className={classes.media}
-//         image="../images/Restaurant.jpg"
-//         title="Paella dish"
-//       />
-//       <CardContent>
-//         <Typography variant="body2" color="textSecondary" component="p">
-//           Booking for a lovely restaurant
-//         </Typography>
-//       </CardContent>
-//       <CardActions disableSpacing>
-//         <IconButton aria-label="Add to favorites">
-//           <FavoriteIcon />
-//         </IconButton>
-//         <IconButton aria-label="Share">
-//           <ShareIcon />
-//         </IconButton>
-//         <IconButton
-//           className={clsx(classes.expand, {
-//             [classes.expandOpen]: expanded
-//           })}
-//           onClick={handleExpandClick}
-//           aria-expanded={expanded}
-//           aria-label="Show more"
-//         >
-//           <ExpandMoreIcon />
-//         </IconButton>
-//       </CardActions>
-//       <Collapse in={expanded} timeout="auto" unmountOnExit>
-//         <CardContent>
-//           <Typography paragraph>Method:</Typography>
-//           <Typography paragraph>
-//             Heat 1/2 cup of the broth in a pot until simmering, add saffron and
-//             set aside for 10 minutes.
-//           </Typography>
-//           <Typography paragraph>
-//             Heat oil in a (14- to 16-inch) paella pan or a large, deep skillet
-//             over medium-high heat. Add chicken, shrimp and chorizo, and cook,
-//             stirring occasionally until lightly browned, 6 to 8 minutes.
-//             Transfer shrimp to a large plate and set aside, leaving chicken and
-//             chorizo in the pan. Add pimentón, bay leaves, garlic, tomatoes,
-//             onion, salt and pepper, and cook, stirring often until thickened and
-//             fragrant, about 10 minutes. Add saffron broth and remaining 4 1/2
-//             cups chicken broth; bring to a boil.
-//           </Typography>
-//           <Typography paragraph>
-//             Add rice and stir very gently to distribute. Top with artichokes and
-//             peppers, and cook without stirring, until most of the liquid is
-//             absorbed, 15 to 18 minutes. Reduce heat to medium-low, add reserved
-//             shrimp and mussels, tucking them down into the rice, and cook again
-//             without stirring, until mussels have opened and rice is just tender,
-//             5 to 7 minutes more. (Discard any mussels that don’t open.)
-//           </Typography>
-//           <Typography>
-//             Set aside off of the heat to let rest for 10 minutes, and then
-//             serve.
-//           </Typography>
-//         </CardContent>
-//       </Collapse>
-//     </Card>
-//   );
-// }
-
-
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import clsx from "clsx";
@@ -147,6 +18,7 @@ import Collapse from "@material-ui/core/Collapse";
 import Dashboard from "./Dashboard";
 import { fetchUserInfo } from '../services/api'
 import moment from "moment"
+import swal from 'sweetalert';
 
 
 const useStyles = makeStyles(theme => ({
@@ -182,12 +54,28 @@ export default function BookingCard(props) {
   }
 
   const handleDelete = id => {
-    debugger
     let my_saved_restaurant = props.savedRestaurants.find(
       restaurant => restaurant.restaurant_api_id === id
     );
-    props.deleteCallback(props.user.id, my_saved_restaurant.id)
-      .then(() => props.secondCallback(props.user));
+    swal({
+      title: "Are you sure you want to cancel your booking?",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          props.deleteCallback(props.user.id, my_saved_restaurant.id)
+          swal("Your booking has been cancelled.", {
+            icon: "success",
+          })
+            .then(() => props.secondCallback(props.user));
+
+        } else {
+          swal("Your booking is still valid.");
+        }
+      });
+
     // return displayDateAndTime()
   };
 
@@ -207,10 +95,20 @@ export default function BookingCard(props) {
     return `${date} @ ${booking_time.time.substring(0, 5)}`
   }
 
-  function imgLoadError(event) {
-    event.target.src = "https://images.unsplash.com/photo-1544148103-0773bf10d330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1050&q=80"
-  }
+  const images = [
+    "https://images.unsplash.com/photo-1544148103-0773bf10d330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1050&q=80",
+    "https://images.unsplash.com/photo-1537047902294-62a40c20a6ae?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=675&q=80",
+    "https://images.unsplash.com/photo-1551218372-a8789b81b253?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=634&q=80",
+    "https://images.unsplash.com/photo-1506812779316-934cef283429?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=634&q=80",
+    "https://images.unsplash.com/photo-1483274816418-3975509c8f78?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80",
+    "https://images.unsplash.com/photo-1483648969698-5e7dcaa3444f?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1334&q=80"
+  ]
 
+  function imgLoadError(event) {
+    event.target.src = images[Math.floor(Math.random() * images.length)]
+
+    // event.target.src = "https://images.unsplash.com/photo-1544148103-0773bf10d330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1050&q=80"
+  }
 
 
   return (
